requirements.txt
python-dotenv
fastapi
SQLAlchemy
psycopg2-binary
alembic
